refactoring

+++ src/FinanceManager.Application/Reports/Commands/RemoveReport/RemoveReportCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using FinanceManager.Application.Common.Exceptions;
 using FinanceManager.Application.Common.Interfaces;
+using FinanceManager.Application.Common.Models;
 using FinanceManager.Domain.Entities;
 using MediatR;
 using System;
@@ -10,7 +11,7 @@ using System.Threading.Tasks;
 
 namespace FinanceManager.Application.Reports.Commands.RemoveReport
 {
-    public class RemoveReportCommandHandler : IRequestHandler<RemoveReportCommand>
+    public class RemoveReportCommandHandler : IRequestHandler<RemoveReportCommand, Result>
     {
         private readonly IReportRepository _reportRepository;
 
@@ -19,18 +20,22 @@ namespace FinanceManager.Application.Reports.Commands.RemoveReport
             _reportRepository = reportRepository;
         }
 
-        public async Task<Unit> Handle(RemoveReportCommand request, CancellationToken cancellationToken)
+        public async Task<Result> Handle(RemoveReportCommand request, CancellationToken cancellationToken)
         {
             var report = await _reportRepository.GetReportByIdAsync(request.ReportId);
 
-            if(report == null)
+            if (report == null)
             {
                 throw new NotFoundException(nameof(Report), request.ReportId);
+            }
+            if (report.DailyReport.AppUserId != request.AppUserId)
+            {
+                return Result.Failure(new string[] { "Неверный id отчета" });
             }
 
             await _reportRepository.RemoveReportAsync(report);
 
-            return Unit.Value; 
+            return Result.Success(); 
         }
     }
 }
